+++ Frontend/src/Sites/LoginComponent.js
@@ -8,7 +8,7 @@ import {
     InputLabel,
     OutlinedInput, InputAdornment, IconButton, FormControl, Alert
 } from "@mui/material";
-import Footer from "./Footer.js";
+import Footer from "../components/Footer.js";
 import {centeredDivStyle, FrameStyle, InputFieldStyle} from "./RegisterPage.js"
 import {Visibility, VisibilityOff} from "@mui/icons-material";
 import * as React from "react";
@@ -16,7 +16,7 @@ import {JWT_TOKEN, ReturnHomeWhenLoggedIn} from "../helper/BackendHelper.js";
 import {useNavigate} from "react-router-dom";
 import FetchBackend from "../helper/BackendHelper.js";
 import Cookies from "js-cookie";
-import Appbar from "./Appbar.js";
+import Appbar from "../components/Appbar.js";
 
 
 function LoginComponent(){
